Move some methods and generic classes to api to make them re-usable
